requirements.txt
fastapi
uvicorn
requests
face_recognition
firebase-admin
python-multipart
python-dotenv
pillow
requests
numpy
opencv-python